Refactoring the file structure of the project.
