+++ app/page.tsx
@@ -6,7 +6,7 @@ export default function Home() {
 
         <div className="container-fluid page ml-1">
         <DirectorInfo/>
-            <div  className="ml-2"><Heading>Our Schemes</Heading></div>
+        <div className="ml-2"><Heading>Our Schemes</Heading></div>
         </div>
